+++ src/main/java/com/example/springbootapi/repo/GoalRepository.java
@@ -3,4 +3,4 @@ package com.example.springbootapi.repo;
 import com.example.springbootapi.model.Goal;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GoalRepository extends JpaRepository<Goal, Integer> {}
+public interface GoalRepository extends JpaRepository<Goal, Long> {}
